requirements.txt
dnspython==2.2.1
flask-pydantic-spec==0.3.1
flask-restful==0.3.9
pip-chill==1.0.1
pymongo==4.0.2
python-decouple==3.6
redis==4.3.0
